quickstart: fix incorrect usage of English English
The quickstart generator appears to use "cheerio" as a synonym for
"hello" or "welcome". This is incorrect in the Southern English dialect
insomuch as "cheerio" means specifically "goodbye" or "so long".

I suggest changing this either explicitly to a more traditional "hello"
or "welcome" message or in the alternative using a more correct idiom
such as "wotcha", "hallo" or "hey nonny nonny".