Remove the forced LF line ending in editorconfig
Forcing LF via `.editorconfig` is super annoying when working on Windows. The line ending normalization in the repo is already forced by `.gitattributes`, the editorconfig one just completely break the git diff for Windows user and it forces us to resave file each time we open it up (as Git will checkout RN on windows). So I'd vote to remove it - please!