Breaking: Remove callbacks/promises from install APIs (Fix #1077)
And rename runInstall to scheduleInstallTask

This work is done to remove a very common area where new Yeoman users
get stuck. The install tasks do not actually run the installation, it
only schedule the installation to run later in the run loop.

This caused a lot of issues with users creating dead-loops in their
generators, and it was really not an obvious error to fix.

The proper way to handle actions after the installation process if to
add a hook into the `end` phase.